Fixed exception

+++ plugins/amoeba/platforms/common/src/AmoebaCommonKernels.cpp
@@ -1255,7 +1255,7 @@ void CommonCalcAmoebaMultipoleForceKernel::computeInducedField() {
     computeInducedFieldKernel->setArg(7, numTileIndices);
     if (usePME) {
         setPeriodicBoxArgs(cc, computeInducedFieldKernel, 10);
-        computeInducedFieldKernel->setArg(15, nb.getInteractingTiles().getSize());
+        computeInducedFieldKernel->setArg(15, (int) nb.getInteractingTiles().getSize());
     }
     cc.clearBuffer(inducedField);
     cc.clearBuffer(inducedFieldPolar);
